feat: refactor resource imports and add compositions and printer utilities
