+++ internal/resources/printer.go
