+++ internal/resources/table_printer.go
