+++ internal/util/cmd/flags.go
@@ -1,4 +1,4 @@
-package helpers
+package cmd
 
 import "strings"
